+++ DiscordMusic.Core/Discord/Music/Lyrics/ILyricsService.cs
@@ -1,0 +1,6 @@
+﻿namespace DiscordMusic.Core.Discord.Music.Lyrics;
+
+public interface ILyricsService
+{
+    Task<Lyrics?> GetLyricsAsync(string title, string author);
+}