+++ DiscordMusic.Core/Discord/Music/MusicModule.cs
@@ -1,4 +1,5 @@
 using DiscordMusic.Core.Discord.Music.Download;
+using DiscordMusic.Core.Discord.Music.Lyrics;
 using DiscordMusic.Core.Discord.Music.Queue;
 using DiscordMusic.Core.Discord.Music.Store;
 using DiscordMusic.Core.Discord.Music.Streaming;
@@ -14,6 +15,7 @@ internal static class MusicModule
         services.AddSingleton<IMusicStore, MusicStore>();
         services.AddSingleton<IMusicStreamer, MusicStreamer>();
         services.AddTransient<IMusicDownloader, MusicDownloader>();
+        services.AddTransient<ILyricsService, LyricsService>();
 
         services.AddHostedService<MusicHost>();
     }